backend/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-openai
numpy
sse-starlette
langchain-community
